+++ src/expression/function/Func.ts
@@ -4,7 +4,6 @@ import { isNullish } from "../../util/IfcExpressionUtils.js";
 import { FuncArg } from "./FuncArg.js";
 import {
   ExprEvalError,
-  ExprEvalFunctionEvaluationConsequentialErrorObj,
   ExprEvalMissingRequiredFunctionArgumentErrorObj,
   ExprEvalResult,
   ExprEvalSuccessObj,
@@ -18,6 +17,7 @@ import { ExprType } from "../../type/ExprType.js";
 import { ParserRuleContext } from "antlr4";
 import { FunctionExpr } from "./FunctionExpr.js";
 import { Expr } from "../Expr";
+import { SpuriousFunctionArgumentException } from "../../error/SpuriousFunctionArgumentException";
 
 export abstract class Func {
   protected name: string;
@@ -50,6 +50,15 @@ export abstract class Func {
   ): void {
     const numProvided = providedArgumentTypes.length;
     if (!isNullish(this.formalArguments)) {
+      if (numProvided > this.formalArguments.length) {
+        throw new SpuriousFunctionArgumentException(
+          this.name,
+          "[unexpected argument]",
+          this.formalArguments.length,
+          ctx,
+          `Function expects (at most) ${this.formalArguments.length} arguments`
+        );
+      }
       for (let i = 0; i < this.formalArguments.length; i++) {
         const currentArg: FuncArg<unknown> = this.formalArguments[i];
         if (numProvided > i) {