+++ src/expression/function/impl/AttributeAccessorFunction.ts
@@ -5,7 +5,6 @@ import { ExprEvalResult, ExprEvalSuccessObj } from "../../ExprEvalResult.js";
 import { ObjectAccessor } from "../../../context/ObjectAccessor.js";
 import { ExprType } from "../../../type/ExprType.js";
 import { FunctionExpr } from "../FunctionExpr.js";
-import { ParserRuleContext } from "antlr4";
 
 export class AttributeAccessorFunction extends Func {
   private readonly attributeName;